Move util classes to respective packages
